book_recommendations/requirements.txt
Django==4.2.6
django-bootstrap-v5==1.0.11
django-cors-headers==4.2.0
django-crispy-forms==2.1
django-filter==23.5
django-rest-framework==0.1.0
djangorestframework==3.14.0
entrypoints==0.4
environ==1.0
Flask==2.2.2
Jinja2==3.1.2
numpy==1.23.4
oauthlib==3.2.2
opencv-python==4.8.1.78
PyJWT==2.8.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
sqlparse==0.4.2
virtualenv==20.14.1
virtualenvwrapper-win==1.2.7
